Rename `experiment_search.py` to `search.py`.
