+++ nupic/research/frameworks/vernon/search.py
